+++ SI.Rosetta.Aggregates/IAggregateState.fs
@@ -2,4 +2,4 @@ namespace SI.Rosetta.Aggregates
 
 type IAggregateState =
     abstract member Id: string
-    abstract member Version: int
+    abstract member Version: int64
